requirements.txt
Flask==2.1.3
sling
PyYAML==6.0
Werkzeug==2.3.7
